feat(accounts,services): re-land profile editing + dedupe UI on device-first main (port of #530)
Ports the profile-editing unblock, dead-link fixes, and god-screen
decomposition from PR #530 onto post-cutover main. #530 was merged into
design/cross-domain-session-sync and never reached main; the device-first
cutover (#524-#529) replaced that line, so this re-lands #530's content on top
of the current main reality.

Ported:
- accounts: Edit-profile hub wiring; god-screen decomposition (activity,
  sharing, storage, managed-accounts) into components/hooks/utils; real
  third-party-connections family screen (redundant connected-apps twin deleted);
  phantom location field removed; i18n backfill/prune.
- services: profile-editing feature — EditProfileScreen hub, EditProfileFieldScreen
  fieldType fix + inline display-name validation, useProfileEditing, ProfileScreen,
  ManageAccountScreen, EditProfile routes.
- core: isValidDisplayName (client-side display-name policy), phantom User.location
  removed, editProfile i18n keys.
- contracts: drop phantom location field.

Folded-in review findings (originally applied only on the phantom branch):
- family.tsx retry key connectedApps.retry -> common.retry
- activity-format.ts invalid-date guard (Number.isNaN(getTime()) -> '')
- EditProfileFieldScreen link-title crash guard via linkFormat.ts (String(item ?? ''))
- core isValidDisplayName uses \p{Zs} not \s (rejects tab/newline/CR; matches the
  api displayNameSanitize probe in PR #532 — one policy, no conflicting probes)

Dropped hunks (fight the device-first cutover — main is truth for auth/session):
- account-switcher dedup (delete AccountMenu/ProfileMenu, rewrite AccountSwitcher/
  ProfileButton/exports) — main's post-cutover switcher kept intact.
- inbox MailboxDrawer AccountMenu->AccountSwitcher swap (depends on the dropped dedup).
- api CORS/loopback-origin change + tests — already on main (a966cb4c).

Gates: core 867/867, services 198/198 + build, accounts tsc 117 (2 fewer than the
119 baseline, zero new), inbox unchanged, biome clean on changed files.
