+++ packages/accounts/app/(tabs)/devices/_layout.tsx
@@ -1,6 +1,8 @@
 import { Stack } from 'expo-router';
+import { useTranslation } from '@/lib/i18n';
 
 export default function DevicesLayout() {
+    const { t } = useTranslation();
     return (
         <Stack>
             <Stack.Screen
@@ -13,7 +15,7 @@ export default function DevicesLayout() {
                 name="[deviceId]"
                 options={{
                     presentation: 'modal',
-                    title: 'Device Details',
+                    title: t('devices.detail.title'),
                     headerShown: false,
                 }}
             />
