+++ packages/accounts/app/(tabs)/managed-accounts.tsx
@@ -1,59 +1,20 @@
-import React, { useMemo, useCallback, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { View, StyleSheet, Text, TouchableOpacity, ActivityIndicator } from 'react-native';
-import { useColors, type AppColors } from '@/hooks/useColors';
+import { useColors } from '@/hooks/useColors';
 import { ThemedText } from '@/components/themed-text';
 import { ScreenContentWrapper } from '@/components/screen-content-wrapper';
-import { ScreenHeader, AccountCard } from '@/components/ui';
+import { ScreenHeader, AccountCard, EmptyStateCard } from '@/components/ui';
 import { Section } from '@/components/section';
 import { GroupedSection } from '@/components/grouped-section';
-import type { GroupedItem } from '@/components/sections/types';
-import { useOxy, Avatar } from '@oxyhq/services';
+import { useOxy } from '@oxyhq/services';
 import { alert, toast } from '@oxyhq/bloom';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { useHapticPress } from '@/hooks/use-haptic-press';
 import { useTranslation } from '@/lib/i18n';
-import type { AccountNode, AccountRole } from '@oxyhq/core';
-import { getAccountDisplayName as coreGetAccountDisplayName, getAccountFallbackHandle } from '@oxyhq/core';
-
-// Roles whose membership carries the `account:act_as` capability — the only
-// roles that can SWITCH INTO the account (selecting it makes the whole app
-// become that account). Mirrors the account role set (owner/admin/editor).
-const SWITCHABLE_ROLES: readonly AccountRole[] = ['owner', 'admin', 'editor'];
-// Roles that may manage membership + sharing of the account.
-const MANAGE_MEMBER_ROLES: readonly AccountRole[] = ['owner', 'admin'];
-// Roles that may edit the account's profile.
-const EDIT_ROLES: readonly AccountRole[] = ['owner', 'admin', 'editor'];
-
-function getRoleBadgeColor(role: AccountRole, colors: AppColors): string {
-  switch (role) {
-    case 'owner':
-      return colors.sidebarIconPersonalInfo;
-    case 'admin':
-      return colors.sidebarIconSecurity;
-    case 'editor':
-      return colors.sidebarIconPayments;
-    case 'developer':
-      return colors.sidebarIconData;
-    case 'billing':
-      return colors.sidebarIconStorage;
-    default:
-      return colors.icon;
-  }
-}
-
-function getAccountDisplayName(node: AccountNode, locale?: string): string {
-  return coreGetAccountDisplayName(node.account ?? null, locale);
-}
-
-/**
- * The caller's effective role on a node. Prefer the resolved membership role;
- * fall back to the relationship (an owned/`self` account is implicitly owned,
- * a shared account with no membership row defaults to read-only `viewer`).
- */
-function getNodeRole(node: AccountNode): AccountRole {
-  if (node.callerMembership?.role) return node.callerMembership.role;
-  return node.relationship === 'member' ? 'viewer' : 'owner';
-}
+import type { AccountNode } from '@oxyhq/core';
+import { getDisplayName } from '@/utils/date-utils';
+import { useAccountRowBuilder } from '@/components/managed-accounts/account-row';
+import { useManagedAccountGroups } from '@/hooks/managed-accounts/useManagedAccountGroups';
 
 export default function ManagedAccountsScreen() {
   const colors = useColors();
@@ -123,7 +84,7 @@ export default function ManagedAccountsScreen() {
   }, [switchToAccount, showBottomSheet]);
 
   const handleArchiveAccount = useCallback((node: AccountNode) => {
-    const name = getAccountDisplayName(node, locale);
+    const name = getDisplayName(node.account ?? null, locale);
     alert(
       t('managedAccounts.archive.confirmTitle'),
       t('managedAccounts.archive.confirmBody', { name }),
@@ -149,136 +110,17 @@ export default function ManagedAccountsScreen() {
     );
   }, [oxyServices, refreshAccounts, locale, t]);
 
-  const buildItem = useCallback((node: AccountNode): GroupedItem => {
-    const name = getAccountDisplayName(node, locale);
-    const username = node.account?.username;
-    // When an account has no username yet (e.g. mid-provisioning) fall back to
-    // a truncated `publicKey` handle so the row still reads as identifiable
-    // rather than showing "No username set".
-    const fallbackHandle = getAccountFallbackHandle(node.account ?? null);
-    const role = getNodeRole(node);
-    // "Current" = the account the app is currently signed in as. A real session
-    // switch makes `user` become the switched-into account, so the current row
-    // is simply the one whose `accountId` matches `user.id`. The personal/self
-    // account is not listed here.
-    const isCurrent = user?.id === node.accountId;
-    const isArchiving = archivingId === node.accountId;
-    const avatarUri = node.account?.avatar
-      ? oxyServices.getFileDownloadUrl(node.account.avatar, 'thumb')
-      : undefined;
-    const badgeColor = getRoleBadgeColor(role, colors);
-    const canSwitchInto = SWITCHABLE_ROLES.includes(role);
-    const canManageMembers = MANAGE_MEMBER_ROLES.includes(role);
-    const canEdit = EDIT_ROLES.includes(role);
-    const canArchive = role === 'owner';
+  const buildItem = useAccountRowBuilder({
+    currentAccountId: user?.id ?? null,
+    archivingId,
+    oxyServices,
+    onSwitchTo: handleSwitchTo,
+    onManageMembers: handleManageMembers,
+    onEditProfile: handleEditProfile,
+    onArchive: handleArchiveAccount,
+  });
 
-    return {
-      id: node.accountId,
-      title: name,
-      subtitle: username
-        ? `@${username}`
-        : (fallbackHandle ?? t('managedAccounts.noUsernameYet')),
-      customIcon: <Avatar name={name} uri={avatarUri} size={40} />,
-      customContent: (
-        <View style={styles.accountActions}>
-          <View style={[styles.roleBadge, { backgroundColor: badgeColor + '20', borderColor: badgeColor + '40' }]}>
-            <Text style={[styles.roleBadgeText, { color: badgeColor }]}>{role}</Text>
-          </View>
-          {isArchiving ? (
-            <ActivityIndicator size="small" color={colors.text} />
-          ) : (
-            <View style={styles.actionButtons}>
-              {isCurrent ? (
-                // Single "current account" indicator (Gmail-style). Not a toggle:
-                // the account is already active, so there is nothing to press.
-                <View
-                  style={styles.currentIndicator}
-                  accessible
-                  accessibilityRole="image"
-                  accessibilityLabel={t('a11y.stopActingAs')}
-                >
-                  <MaterialCommunityIcons name="check" size={20} color={colors.tint} />
-                </View>
-              ) : canSwitchInto ? (
-                <TouchableOpacity
-                  style={[styles.actionButton, { backgroundColor: colors.card }]}
-                  onPressIn={handlePressIn}
-                  onPress={() => handleSwitchTo(node.accountId)}
-                  activeOpacity={0.7}
-                  accessibilityRole="button"
-                  accessibilityLabel={t('a11y.actAs')}
-                >
-                  <MaterialCommunityIcons name="swap-horizontal" size={16} color={colors.text} />
-                </TouchableOpacity>
-              ) : null}
-              {canManageMembers && (
-                <TouchableOpacity
-                  style={[styles.actionButton, { backgroundColor: colors.card }]}
-                  onPressIn={handlePressIn}
-                  onPress={() => handleManageMembers(node.accountId)}
-                  activeOpacity={0.7}
-                  accessibilityRole="button"
-                  accessibilityLabel={t('a11y.manageMembers')}
-                >
-                  <MaterialCommunityIcons name="account-multiple-outline" size={16} color={colors.text} />
-                </TouchableOpacity>
-              )}
-              {canEdit && (
-                <TouchableOpacity
-                  style={[styles.actionButton, { backgroundColor: colors.card }]}
-                  onPressIn={handlePressIn}
-                  onPress={() => handleEditProfile(node.accountId)}
-                  activeOpacity={0.7}
-                  accessibilityRole="button"
-                  accessibilityLabel={t('a11y.editProfile')}
-                >
-                  <MaterialCommunityIcons name="pencil-outline" size={16} color={colors.text} />
-                </TouchableOpacity>
-              )}
-              {canArchive && (
-                <TouchableOpacity
-                  style={[styles.actionButton, { backgroundColor: colors.card }]}
-                  onPressIn={handlePressIn}
-                  onPress={() => handleArchiveAccount(node)}
-                  activeOpacity={0.7}
-                  accessibilityRole="button"
-                  accessibilityLabel={t('a11y.archiveAccount')}
-                >
-                  <MaterialCommunityIcons name="archive-outline" size={16} color={colors.error} />
-                </TouchableOpacity>
-              )}
-            </View>
-          )}
-        </View>
-      ),
-      // Tapping the row switches INTO the account (Gmail-style). The current
-      // account is not pressable (it is already active); accounts the caller
-      // can't switch into fall back to the manage-members affordance.
-      onPress: isCurrent
-        ? undefined
-        : canSwitchInto
-          ? () => handleSwitchTo(node.accountId)
-          : canManageMembers
-            ? () => handleManageMembers(node.accountId)
-            : undefined,
-    };
-  }, [user?.id, archivingId, oxyServices, colors, handlePressIn, handleSwitchTo, handleManageMembers, handleEditProfile, handleArchiveAccount, t, locale]);
-
-  // Partition the accessible forest: accounts the caller owns (grouped by kind)
-  // and accounts shared with them via membership. The caller's own personal
-  // (`self`) account is naturally excluded — it is neither a managed kind nor a
-  // `member` relationship.
-  const groups = useMemo(() => {
-    const owned = accounts.filter((a) => a.relationship !== 'member' && a.kind !== 'personal');
-    return {
-      organizations: owned.filter((a) => a.kind === 'organization'),
-      projects: owned.filter((a) => a.kind === 'project'),
-      bots: owned.filter((a) => a.kind === 'bot'),
-      shared: accounts.filter((a) => a.relationship === 'member'),
-    };
-  }, [accounts]);
-
-  const totalCount = groups.organizations.length + groups.projects.length + groups.bots.length + groups.shared.length;
+  const { groups, totalCount } = useManagedAccountGroups(accounts);
 
   if (oxyLoading) {
     return (
@@ -316,15 +158,11 @@ export default function ManagedAccountsScreen() {
 
           {totalCount === 0 ? (
             <Section title="">
-              <View style={styles.emptyState}>
-                <View style={[styles.emptyIcon, { backgroundColor: colors.sidebarIconSharing + '20' }]}>
-                  <MaterialCommunityIcons name="account-group-outline" size={48} color={colors.sidebarIconSharing} />
-                </View>
-                <ThemedText style={styles.emptyTitle}>{t('managedAccounts.empty.title')}</ThemedText>
-                <ThemedText style={[styles.emptySubtitle, { color: colors.icon }]}>
-                  {t('managedAccounts.empty.subtitle')}
-                </ThemedText>
-              </View>
+              <EmptyStateCard
+                icon="account-group-outline"
+                title={t('managedAccounts.empty.title')}
+                subtitle={t('managedAccounts.empty.subtitle')}
+              />
             </Section>
           ) : (
             <>
@@ -394,64 +232,5 @@ const styles = StyleSheet.create({
     color: '#FFFFFF',
     fontSize: 16,
     fontWeight: '600',
-  },
-  accountActions: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-  },
-  actionButtons: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 6,
-  },
-  currentIndicator: {
-    width: 32,
-    height: 32,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  actionButton: {
-    width: 32,
-    height: 32,
-    borderRadius: 16,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  roleBadge: {
-    paddingHorizontal: 8,
-    paddingVertical: 3,
-    borderRadius: 10,
-    borderWidth: 1,
-  },
-  roleBadgeText: {
-    fontSize: 11,
-    fontWeight: '600',
-    textTransform: 'capitalize',
-  },
-  emptyState: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingVertical: 48,
-    paddingHorizontal: 32,
-  },
-  emptyIcon: {
-    width: 96,
-    height: 96,
-    borderRadius: 48,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginBottom: 24,
-  },
-  emptyTitle: {
-    fontSize: 20,
-    fontWeight: '600',
-    marginBottom: 8,
-    textAlign: 'center',
-  },
-  emptySubtitle: {
-    fontSize: 15,
-    lineHeight: 22,
-    textAlign: 'center',
   },
 });
