+++ packages/core/src/index.ts
@@ -443,6 +443,7 @@ export {
     isValidEmail,
     isValidUsername,
     isValidPassword,
+    isValidDisplayName,
     isRequiredString,
     isRequiredNumber,
     isRequiredBoolean,
