+++ packages/core/src/models/interfaces.ts
@@ -106,7 +106,6 @@ export interface User {
   phone?: string;
   address?: string;
   birthday?: string;
-  location?: string;
   website?: string;
   createdAt?: string;
   updatedAt?: string;
